Add settings XML serialization

+++ FilmsToWatch/Weather/Wind.cs
@@ -2,7 +2,7 @@
 {
     public class Wind
     {
-        public int speed { get; set; }
+        public double speed { get; set; }
         public int deg { get; set; }
     }
 }
